fix(yt:shared): moved getNatureFromURL from backend
